Relocate var
Relocate var

The commit relocates the variable $settings_cloudflare__zone_api_token to write_record function. The function now accepts an additional parameter for the token, which is used in place of the previous variable. This change ensures that the correct token is passed to Cloudflare's API when updating DNS records.